Multiplexor sprite priority based on Y position
Added Multiplex_MaintainPriority to the multiplexor.

Added this option to the Scroller/SEUCK engine. It's disabled by default
because SEUCK games use of this feature.

This improves the sprite priority calculation for games like Commando so
that the sprites appear with the correct priority based on the Y
position.

Apparently some people thought this would be too hard to do in a
"Commando" type game. So I enabled the second SEUCK demo called "Outlaw"
to demonstrate it can be easily done in that type of game.