Random running enemy AI with background collision
Enable EnemyAITest to allow the enemy AI test code to be enabled.
This will replace the all enemy object type 24 with an enemy that will
run around. (.doRunningMan in AnimationEngine.a)
Enable the ENEMYTST.P00 in ScrollerData.a to include a suitable game
file.
This example file contains a first screen (for 60 seconds) with a "box"
shape where three AI units will try to escape.
They will be constrained by this box until it is removed by the next
screen.