Enemies can now shoot at the player. (EnemyAimTest)
Enable EnemyAimTest to allow enable one enemy that will use directional
fire instead of random fire.
This will replace the all enemy object type 23 with an enemy that will
run around. (.doAimedFire in AnimationEngine.a)
Enable the NMYFRTST.P00 in ScrollerData.a to include a suitable game
file.
This example file is a first screen (for 60 seconds) with an enemy that
circles the player.
The enemy will shoot in the general direction of the player even though
it is set to random in SEUCK.