Tusari 2 - The River Demo V4
It's the first C64 game to use vertical parallax and horizontal parallax with variable speed while using simultaneous raster synchronised NMI and IRQ interrupts for the scroll and sprites. It's also, I think, the first game to use a new technique called Sprite Shift-Plexing which is used for the bullets and the smartbomb effect. All of the bullets on screen only use one sprite, but it is shifted horizontally so each sprite can contain up to three bullets and then the same sprite is also moved vertically to fill the screen.